third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-outline-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Tests that the outline shorthand is computed properly.")

var testContainer = document.createElement("div");
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = '<div style="width:100px;height:100px"><div id="test">hello</div></div>';

e = document.getElementById('test');
computedStyle = window.getComputedStyle(e, null);

e.style.outline = "thick solid red";
shouldBe("computedStyle.getPropertyValue('outline')", "'rgb(255, 0, 0) solid 5px'");

e.style.outline = "2em solid";
shouldBe("computedStyle.getPropertyValue('outline')", "'rgb(0, 0, 0) solid 32px'");

e.style.outline = "4px none red";
shouldBe("computedStyle.getPropertyValue('outline')", "'rgb(255, 0, 0) none 4px'");

document.body.removeChild(testContainer);

</script>
</body>
</html>
